refactor: Extracts Obsidian Link/Create logic
From keymaps to its own file under /utils. So keymaps is cleaner
and does not contain business logic
